requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.8
django-allauth==0.42.0
django-cogwheels==0.3
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
l18n==2018.5
M2Crypto==0.36.0
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.7
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.9.3
wagtail-autocomplete==0.6
wagtailtrans==2.1
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.9
zipp==3.1.0